Patch files from gerrit commits
This is patch files from all of the gerrit commits that
were in review december 2019.

These were created because OCF is moving from a gerrit managed
project to gitlab managed project.

There is some concern that not all developers will be prepared
for the move to gitlab and will need access to there work.

Signed-off-by: George Nash <george.nash@intel.com>